requirements.txt
python-telegram-bot>=21.0
PyYAML>=6.0
openpyxl>=3.1
pytest>=8.0
